Fix client-web ESLint config for ESLint 9
Rename eslint.config.jss to eslint.config.js (typo) and convert from
legacy .eslintrc format to ESLint 9 flat config. Remove old .eslintrc.js.
Add globals devDependency. Auto-fix 32 quote-style errors.

Also fix root lint-staged to run eslint within each package's context
via pnpm -C, so the correct ESLint version is used per package.
