+++ client-web/eslint.config.js
@@ -1,0 +1,42 @@
+const globals = require('globals');
+const tseslint = require('@typescript-eslint/eslint-plugin');
+const tsparser = require('@typescript-eslint/parser');
+const prettier = require('eslint-plugin-prettier');
+
+const env = (prod, dev) => (process.env.NODE_ENV === 'production' ? prod : dev);
+
+/** @type {import('eslint').Linter.Config[]} */
+module.exports = [
+  {
+    files: ['**/*.{js,ts}'],
+    languageOptions: {
+      parser: tsparser,
+      globals: {
+        ...globals.node,
+      },
+    },
+    plugins: {
+      '@typescript-eslint': tseslint,
+      prettier,
+    },
+    rules: {
+      quotes: ['error', 'single', { avoidEscape: true }],
+      'no-console': env(1, 0),
+      'no-debugger': env(1, 0),
+      '@typescript-eslint/interface-name-prefix': 'off',
+      '@typescript-eslint/explicit-function-return-type': 'off',
+      '@typescript-eslint/explicit-module-boundary-types': 'off',
+      '@typescript-eslint/no-explicit-any': 'off',
+      '@typescript-eslint/no-unused-vars': [
+        env(2, 1),
+        {
+          argsIgnorePattern: '^_',
+        },
+      ],
+      'no-multiple-empty-lines': 'error',
+    },
+  },
+  {
+    ignores: ['**/node_modules/**', '**/dist/**'],
+  },
+];
