+++ client-web/package.json
@@ -53,6 +53,7 @@
     "dotenv": "^16.0.0",
     "eslint": "^9.17.0",
     "eslint-plugin-prettier": "^5.2.1",
+    "globals": "^15.14.0",
     "graphql-request": "6.0.0",
     "prettier": "^3.4.2",
     "ts-node": "^10.9.2",
