+++ client-web/src/functional-e2e/templates/template-types/usage/contributions/callout-template.use.memos.ts
@@ -1,6 +1,6 @@
-import { Locator, Page, expect } from "@playwright/test";
-import { CalloutTemplateForm } from "../../forms/callout/callout-template-form.models";
-import { getCreateContributionDialog } from "./callout-template.use.contributions";
+import { Locator, Page, expect } from '@playwright/test';
+import { CalloutTemplateForm } from '../../forms/callout/callout-template-form.models';
+import { getCreateContributionDialog } from './callout-template.use.contributions';
 
 export const verifyCalloutContributionMemos = async (
   page: Page,
@@ -31,7 +31,7 @@ export const verifyCalloutContributionMemos = async (
     await dialog.getByRole('button', { name: 'Create' }).click();
 
     // Verify the url has changed to include the new memo
-    await expect(page).toHaveURL(new RegExp(`/memos/`));
+    await expect(page).toHaveURL(new RegExp('/memos/'));
 
     // Verify the contribution appears in the callout
     await expect(
