+++ client-web/src/functional-e2e/templates/template-types/usage/contributions/callout-template.use.whiteboards.ts
@@ -1,6 +1,6 @@
-import { Locator, Page, expect } from "@playwright/test";
-import { CalloutTemplateForm } from "../../forms/callout/callout-template-form.models";
-import { getCreateContributionDialog } from "./callout-template.use.contributions";
+import { Locator, Page, expect } from '@playwright/test';
+import { CalloutTemplateForm } from '../../forms/callout/callout-template-form.models';
+import { getCreateContributionDialog } from './callout-template.use.contributions';
 
 export const verifyCalloutContributionWhiteboards = async (
   page: Page,
@@ -31,7 +31,7 @@ export const verifyCalloutContributionWhiteboards = async (
     await dialog.getByRole('button', { name: 'Create' }).click();
 
     // Verify the url has changed to include the new post
-    await expect(page).toHaveURL(new RegExp(`/whiteboards/`));
+    await expect(page).toHaveURL(new RegExp('/whiteboards/'));
 
     // Verify the contribution appears in the callout
     await expect(
